Use subsurfaces for item instances
This commit adds a subsurface for every item instance. This change greatly
simplifies drawing logic and allows us to render less on average.

In the future, the cursor motion handling could be replaced by enter/leave
events on the subsurfaces, although I am still not entirely sure if that is a
good idea.