Bug (and fix): svg icons regression
Due to changes made after bumping the librsvg version in use, a bug was introduced whereby the viewBox property is now used in place of height and width. However, the viewBox property is not always present in an svg file. Additionally, only the height and width were being validated, and then used to calculate the scale for displaying the image - this should have been updated to use the height and width of the viewbox

Fix: I've included a patch which makes the viewbox height and width calculations more forgiving, and uses the correct values when calculating scale